rename mips module
